Added LuaTransferFun2DProxy class.
This class acts as a proxy for TransferFunction2D in IO. Necessary
because IO is not aware of the LuaScripting system.

+++ LuaScripting/TuvokSpecific/LuaTransferFun1DProxy.cpp
@@ -122,7 +122,7 @@ bool LuaTransferFun1DProxy::proxyLoadWithFilenameAndSize(
 
 //------------------------------------------------------------------------------
 void LuaTransferFun1DProxy::proxySetStdFunction(
-    float centerPoint, float invGradient, float component, bool invertedStep)
+    float centerPoint, float invGradient, int component, bool invertedStep)
 {
   if (m1DTrans == NULL) return;
   m1DTrans->SetStdFunction(centerPoint, invGradient, component, invertedStep);
